biblioteka v1.2: - przeniesiono enum Options do klasy LibraryControl
