+++ src/pl/javastart/library/app/LibraryApp.java
@@ -2,7 +2,7 @@ package pl.javastart.library.app;
 
 public class LibraryApp {
 
-    private static final String APP_NAME = "Biblioteka v1.1";
+    private static final String APP_NAME = "Biblioteka v1.2";
 
     public static void main(String[] args) {
         System.out.println(APP_NAME);
