+++ src/pl/javastart/library/model/Publication.java
@@ -2,7 +2,7 @@ package pl.javastart.library.model;
 
 import java.util.Objects;
 
-public class Publication {
+public abstract class Publication {
     private String title;
     private String publisher;
     private int year;
